backend/requirements.txt
Flask==2.3.2
Flask-SQLAlchemy==3.0.3
psycopg2-binary==2.9.6
Flask-Migrate==4.0.4
Flask-Cors==4.0.0
Flask-Script==2.0.6
python-dotenv==1.0.1
flask-jwt-extended==4.6.0
